Replace live datashader examples with GIFs
